+++ test/prom_ex/plugins/broadway_test.exs
@@ -1,0 +1,27 @@
+defmodule PromEx.Plugins.BroadwayTest do
+  use ExUnit.Case, async: true
+
+  alias PromEx.Plugins.Broadway
+  alias PromEx.Test.Support.{Events, Metrics}
+
+  defmodule WebApp.PromEx do
+    use PromEx, otp_app: :web_app
+
+    @impl true
+    def plugins do
+      [Broadway]
+    end
+  end
+
+  test "telemetry events are accumulated" do
+    start_supervised!(WebApp.PromEx)
+    Events.execute_all(:broadway)
+
+    metrics =
+      WebApp.PromEx
+      |> PromEx.get_metrics()
+      |> Metrics.sort()
+
+    assert metrics == Metrics.read_expected(:broadway)
+  end
+end